NativeLoader loads JNA libraries
