+++ build.gradle
@@ -123,7 +123,7 @@ tasks.withType(Jar) {
 
 // Build tasks
 
-task buildAltbn128(group: 'build', description: 'builds altbn128 library with system go') {
+task buildAltbn128(group: 'build', description: 'builds altbn128 library') {
     doLast {
         exec {
             workingDir 'altbn128'
@@ -132,7 +132,7 @@ task buildAltbn128(group: 'build', description: 'builds altbn128 library with sy
     }
 }
 
-task buildSecp256k1(group: 'build', description: 'builds secp256k1 library for current OS (helpful for development)') {
+task buildSecp256k1(group: 'build', description: 'builds secp256k1 library') {
     dependsOn 'generateSecp256k1JniHeaders'
     dependsOn 'replaceSecp256k1JniHeaders'
     doLast {
@@ -143,7 +143,7 @@ task buildSecp256k1(group: 'build', description: 'builds secp256k1 library for c
     }
 }
 
-task buildBls12_381(group: 'build', description: 'builds BLS12_381 library for current OS') {
+task buildBls12_381(group: 'build', description: 'builds BLS12_381 library') {
     doLast {
         exec {
             workingDir 'bls12_381'
