+++ src/main/java/co/rsk/altbn128/JniBn128.java
@@ -1,5 +1,7 @@
 package co.rsk.altbn128;
 
+import co.rsk.NativeLoader;
+
 public class JniBn128 {
     public native int add(byte[] input, int len, byte[] output);
 
@@ -9,7 +11,7 @@ public class JniBn128 {
 
     static {
         try {
-            if (NativeLoader.load()) {
+            if (NativeLoader.registerJNI(JniBn128.class, "bn128")) {
                 System.out.println("Successfully loaded.");
             }
         } catch (Exception e) {
